# Conflicts:
#	MaterialControls.podspec
#	iOSUILib/iOSUILib/Calendar/MDDatePickerDialog.h
#	iOSUILib/iOSUILib/Calendar/MDTimePickerDialog.h
#	iOSUILib/iOSUILib/Calendar/MDTimePickerDialog.m
#	iOSUILib/iOSUILib/MDButton.h
#	iOSUILib/iOSUILib/MDCollectionViewCell.m
#	iOSUILib/iOSUILib/MDProgress.h
#	iOSUILib/iOSUILib/MDSlider.h
#	iOSUILib/iOSUILib/MDSwitch.h
#	iOSUILib/iOSUILib/MDTabBar.h
#	iOSUILib/iOSUILib/MDTabBar.m
#	iOSUILib/iOSUILib/MDTabBarViewController.h
#	iOSUILib/iOSUILib/MDTabBarViewController.m
#	iOSUILib/iOSUILib/MDTextField.h
#	iOSUILib/iOSUILib/Slider/MDSliderThumbView.h
#	iOSUILibDemo.xcodeproj/xcuserdata/michaellau.xcuserdatad/xcschemes/iOSUILibDemo.xcscheme
#	iOSUILibDemo.xcodeproj/xcuserdata/michaellau.xcuserdatad/xcschemes/xcschememanagement.plist

+++ iOSUILibDemo/TabBarViewControllerViewController.m
@@ -100,7 +100,10 @@
                                              options:0
                                              metrics:nil
                                                views:viewsDictionary]];
+
+  //  tabBarViewController.selectedIndex = 3;
   self.title = @"MDTabBarViewController";
+  
 }
 
 - (void)didReceiveMemoryWarning {
@@ -115,7 +118,7 @@
   TabContentViewController *controller =
       [[TabContentViewController alloc] init];
   dispatch_async(dispatch_get_main_queue(), ^{
-    [controller setContent:[NSString stringWithFormat:@"Tab %i", index + 1]];
+    [controller setContent:[NSString stringWithFormat:@"Tab %d", (int)index + 1]];
   });
   return controller;
 }
